requirements.txt
langchain==1.0.2
langchain-community==0.4
langchain-groq===1.0.0
chromadb==1.2.1
streamlit==1.50.0
pandas== 2.3.3
python-dotenv==1.1.1